standalone vs embedded This PR converts the exporter to a dual mode binary
The previous behaviour is preserved when run with a new argument
`embedded`which reads from stdin to track metrics.

The new mode `standalone` runs external to exabgp and leverages
`exabgpcli` to accomplish scraping.

Due to the differing levels of information available via exabgpcli, the
labels changed a bit so they were unified between modes.

Additionally, the metrics were converted to a more standard prometheus
model with a naming consistent with official exporters.